+++ app/main.js
@@ -44,83 +44,44 @@
 /* 0 */
 /***/ function(module, exports, __webpack_require__) {
 
-	"use strict";
+	/* WEBPACK VAR INJECTION */(function(__dirname) {"use strict";
 
-	var _importExample = __webpack_require__(1);
+	/*  
+	Experimental Websocket Server
+	=============================
 
-	function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } } /*  
-	                                                                                                                                                          Starter JS
-	                                                                                                                                                          ==========
-	                                                                                                                                                          
-	                                                                                                                                                          Starter tempalte for JS projects
-	                                                                                                                                                          
-	                                                                                                                                                          (Shaun A. Noordin || shaunanoordin.com || 20160509)
-	                                                                                                                                                          ********************************************************************************
-	                                                                                                                                                           */
+	Starter tempalte for JS projects
 
-	/*  Primary App Class
+	(Shaun A. Noordin || shaunanoordin.com || 20160509)
+	********************************************************************************
 	 */
-	//==============================================================================
 
-	var App = function App() {
-	  _classCallCheck(this, App);
+	var express = __webpack_require__(1);
+	var server = express();
 
-	  var importExample = new _importExample.ImportExample("HI THERE");
-
-	  this.console = document.getElementById("console");
-	  this.console.innerHTML = "This is a starter template for JS projects. <br>" + importExample.getText();
+	var SERVER = {
+	  PORT: 3000
 	};
-	//==============================================================================
 
-	/*  Initialisations
-	 */
-	//==============================================================================
+	server.set("port", process.env.PORT || SERVER.PORT);
+	server.use(express.static("/web"));
 
+	console.log(__dirname + "/web");
 
-	var app;
-	window.onload = function () {
-	  window.app = new App();
-	};
-	//==============================================================================
+	server.listen(server.get("port"), function onStart(err) {
+	  if (err) {
+	    console.log(err);
+	  } else {
+	    console.info("Server ready on port " + server.get("port"));
+	  }
+	});
+	/* WEBPACK VAR INJECTION */}.call(exports, "/"))
 
 /***/ },
 /* 1 */
 /***/ function(module, exports) {
 
-	"use strict";
-
-	Object.defineProperty(exports, "__esModule", {
-	  value: true
-	});
-
-	var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
-
-	function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
-
-	/*  Example Class
-	    Demonstrates ES6 import function in main.js
-	 */
-	//==============================================================================
-
-	var ImportExample = exports.ImportExample = function () {
-	  function ImportExample() {
-	    var txt = arguments.length <= 0 || arguments[0] === undefined ? "DEFAULT TEXT" : arguments[0];
-
-	    _classCallCheck(this, ImportExample);
-
-	    this.txt = txt;
-	  }
-
-	  _createClass(ImportExample, [{
-	    key: "getText",
-	    value: function getText() {
-	      return "An ImportExample class was imported and has the text: " + this.txt;
-	    }
-	  }]);
-
-	  return ImportExample;
-	}();
-	//==============================================================================
+	module.exports = require("express");
 
 /***/ }
 /******/ ]);
