--- index.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="UTF-8">
-<title>Starter JS</title>
-<link href="./app/main.css" rel="stylesheet">
-<script src="./app/main.js"></script>
-</head>
-<body>
-<div id="app">
-<header>
-  <h1>Starter JS</h1>
-</header>
-<main>
-  <output id="console"></output>
-  <table>
-    <tr>
-      <th>Fruit</th>
-      <th>Colour</th>
-    </tr>
-    <tr>
-      <td><a href="#">Apple</a></td>
-      <td>Red, Green</td>
-    </tr>
-    <tr>
-      <td><a href="#">Banana</a></td>
-      <td>Yellow, Green</td>
-    </tr>
-    <tr>
-      <td><a href="#">Cherry</a></td>
-      <td>Red</td>
-    </tr>
-  </table>
-</main>
-<footer>2016</footer>
-</div><!--/#app-->
-</body>
-</html>
