+++ package.json
@@ -10,20 +10,20 @@
     "presets": ["es2015"]
   },
   "dependencies": {
-    "express": "4.13.4",
+    "express": "4.13.4"
+  },
+  "devDependencies": {
+    "nodemon": "1.10.2",
     "concurrently": "2.0.0",
     "webpack": "1.13.1",
     
     "babel-core": "6.9.1",
     "babel-loader": "6.2.4", 
-    "babel-preset-es2015": "6.9.0",
-    
-    "stylus": "0.54.5"
+    "babel-preset-es2015": "6.9.0"
   },
   "scripts": {
-    "start": "node server.js",
-    "build": "webpack && stylus src -o app",
-    "dev": "concurrently \"webpack --watch\" \"stylus src -o app --watch\" \"node server.js\""
+    "start": "webpack && stylus src -o app && node ./app/main.js",
+    "dev": "concurrently \"webpack --watch\" \"stylus src -o app --watch\" \"nodemon ./app/main.js\""
   },
   "engines": {
     "node": "6.2.0"
