+++ web/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="UTF-8">
+    <title>Experimental Websocket Server</title>
+  </head>
+  <body>
+    <h1>Experimental Websocket Server</h1>
+  </body>
+</html>
